Fixed issue with wall jump
I fixed the issue with the wall jump not being high enough and compounding with the double jump when it was unlocked. There is still a slight issue of not gaining a double jump by simply wall sliding, but it is minor enough to not cause any major difficulties to gameplay. Collectable abilities are also destructible, but can't be gained through an interaction button like I had wanted.