Attempted scene management
Tried to get the player to move between scenes and spawn at certian points. While the code would initially move the player to new scenes, I couldn't get the player to spawn at specific positions. This left most of the code not working, and a different approach is most likely necessary to get the code working.