Level teleports done, camera boundaries need work
Added back in the teleport system, instead of across scenes it does from level to level in the current scene, also setting the player's respawn to the point they spawn in. Need to add a transition to that. Tried adding camera borders to ensure player couldn't see off screen, but the various attempts didn't work and needs to be finished at some point, or at least worked around.