Fixed camera bounds and moving camera
Camera now stops at appropriate bounds using Cinemachine (may have to download on different computers). Also added to the teleport script to be able to move the camera with the player and set new bounds for the camera, allowing easy room by room creation for the game now. Need to make a transition screen to cover a slight jitter with the camera (which is why the script now waits), which will also need to include figuring out how to disable player controls temporarily.