Finished crusher and buzzsaw
Had to fix position errors with both the crusher and the buzzsaw. Added more functionality to the crusher by making it act like the ground on top and walls on the side, along with making a variation where the values switch when it's vertical. Also got some tile sets made for the killzones and one way platforms to help blend into the game better and remove redundant features.