Revised old tile set and added moving platform
Revised my old tile set to get a better quality and connection. Also created a moving platform that stops, has variable distance through animation and acts as a one way platform. Fixed an issue where player would rotate weirdly on said platform. added lights to the crusher to give a more construction feel to the object. Planning to impliment the saw arm next and then move onto a background along with adding further detail to the game. Need to start on collectables and game progression soon.