Added decorations
Added decoration tileset for both the background of the factory and other elements, such as scaffolding and some doors. Still need to sort out an outside background and make a tileset for piles of scrap, along with adding lights for inside the factory. Might add sprite lights to elements that the player can interact with to make it clear to what is and isn't platforms.