Added extra decorations and light
Added scrap decorations as well as lights for the factory. Also added a sprite and idle animation for the main collectables. Need to program the collectables being collected and playing a unique animation before being destroyed. Should be a simple animation trigger on collision enter, followed by a coroutine to wait before destruction on top of adding to the collectable counter.