Finished collectable and added sound system
Fixed all the issues with the collectable, now keeps track of all progress. Was also able to add animations for it. Added a sound effect manager to manage every sound effect in the game and started adding some. Also made it so that you could start to lose hearing of the sound when the player got further away from the source, though it only updates once a sound starts playing, not continuously. Also added music that plays and loops in the level. Need to add sound effects for player then can move onto win conditions, ui and finally level design.