Finished room L
Finished the room, having a one way exit until the player unlocks the dash. Also created a variant of the crusher which the player is unable to stand on top of. Will work on rooms M and N next, which will then give the player the dash and open up level design in the future.