Made room K
Made the room to where the player will run through the room twice if taking this route. I do realise that the path from getting the dash to the double jump is too linear for my liking, so I will add another route to the dash to the right of N. Need to make double jump room, connection to E to the double jump room, and the ending room to have a finished product.