Finished rooms V and W
Have room V done which is the double jump unlock room. This now allows me to design the final levels with the double jump in mind. I also made W, which once other rooms are completed will act as an alternate path to the double jump. I now need to finish the alternate paths for both double jump and dash, which I should also have the time to make extra rooms for secrets. The game is nearly in a finished state.