Finished H and Q
Finished H which completes the second route to the dash. Also finished Q which leads to the hard screen and acts as a second route to the end of the game. Need to finish a few more rooms then work on getting the teleports in place and polishing sound effects.