Added final rooms
Added rooms S and I, which now makes 2 paths to every important location in the game along with the final room. I also added code so that based on the amount of collectables the player has they get sent to different scenes, which will then play different animations. Now I need to make the animations along with a main menu for the game which the player can use to start the game and will get sent back to once the game finishes.