Added ui and menus
Changed the text in the UI to be more fitting to the game. Added a placeholder menu screen which can go into an options menu and into the game. Added an ending cutscene for the normal ending which then sends the player to a results screen. Added a timer to keep track of how long the player has been playing, which encourages replays for better times. Added a results screen which appears after the ending to show the main two stats, collection and time. Plan to add a second ending cutscene, two more rooms to the game, art for the menu and then general polish.