Fixed level transitions
Went around all the current level transitions and fixed the player spawn positions and camera boundaries. Need to fix a camera in room G and S, along with the transitions in rooms V,N, and P. Also added UI for interacting with the ending door and a response for if the player doesn't have enough gears collected. Need to polish a few elements and finish the two ending cutscenes and that's it.